Delete commented-out code, move MockPointerId to tests.rs
and implement PartialEq for PointerId
