+++ core/src/window.rs
@@ -5,7 +5,6 @@ use crate::{
 };
 
 pub trait WindowId: std::fmt::Debug {
-  // fn as_any(&self) -> &dyn Any;
   fn into_any(self: Box<Self>) -> Box<dyn Any>;
   fn equals(&self, other: &Box<dyn WindowId>) -> bool;
   fn box_clone(&self) -> Box<dyn WindowId>;
@@ -13,6 +12,10 @@ pub trait WindowId: std::fmt::Debug {
 
 impl Clone for Box<dyn WindowId> {
   fn clone(&self) -> Self { self.box_clone() }
+}
+
+impl PartialEq for Box<dyn WindowId> {
+  fn eq(&self, other: &Self) -> bool { self.equals(other) }
 }
 
 // pub use winit::window::CursorIcon;
@@ -32,154 +35,6 @@ pub trait RawWindow {
   fn scale_factor(&self) -> f64;
   fn as_any(&self) -> &dyn Any;
 }
-
-// impl RawWindow for winit::window::Window {
-//   fn inner_size(&self) -> Size {
-//     let size = self.inner_size().to_logical(self.scale_factor());
-//     Size::new(size.width, size.height)
-//   }
-
-//   fn set_inner_size(&mut self, size: Size) {
-//     let size = winit::dpi::LogicalSize::new(size.width, size.height);
-//     winit::window::Window::set_inner_size(self, size)
-//   }
-
-//   fn outer_size(&self) -> Size {
-//     let size = self.outer_size().to_logical(self.scale_factor());
-//     Size::new(size.width, size.height)
-//   }
-
-//   fn inner_position(&self) -> Point {
-//     let pos = self
-//       .inner_position()
-//       .expect(" Can only be called on the main thread")
-//       .to_logical(self.scale_factor());
-
-//     Point::new(pos.x, pos.y)
-//   }
-//   #[inline]
-//   fn id(&self) -> WindowId { self.id() }
-
-//   fn outer_position(&self) -> Point {
-//     let pos = self
-//       .outer_position()
-//       .expect(" Can only be called on the main thread")
-//       .to_logical(self.scale_factor());
-//     Point::new(pos.x, pos.y)
-//   }
-
-//   #[inline]
-//   fn request_redraw(&self) { winit::window::Window::request_redraw(self) }
-
-//   fn set_cursor(&mut self, cursor: CursorIcon) { self.set_cursor_icon(cursor)
-// }
-
-//   #[inline]
-//   fn scale_factor(&self) -> f64 { winit::window::Window::scale_factor(self) }
-// }
-
-// pub struct WindowBuilder {
-//   inner_builder: winit::window::WindowBuilder,
-//   root: Widget,
-// }
-
-// impl WindowBuilder {
-//   #[inline]
-//   pub fn build(self, app: &Application) -> Window {
-//     let native_wnd = self.inner_builder.build(app.event_loop()).unwrap();
-//     let size = native_wnd.inner_size();
-//     let ctx = app.context().clone();
-//     let p_backend = AppContext::wait_future(ribir_gpu::wgpu_backend_with_wnd(
-//       &native_wnd,
-//       DeviceSize::new(size.width, size.height),
-//       None,
-//       None,
-//       ctx.shaper.clone(),
-//     ));
-//     Window::new(native_wnd, p_backend, self.root, ctx)
-//   }
-
-//   /// Requests the window to be of specific dimensions.
-//   #[inline]
-//   pub fn with_inner_size(mut self, size: Size) -> Self {
-//     let size = winit::dpi::LogicalSize::new(size.width, size.height);
-//     self.inner_builder = self.inner_builder.with_inner_size(size);
-//     self
-//   }
-
-//   /// Sets a minimum dimension size for the window.
-//   #[inline]
-//   pub fn with_min_inner_size(mut self, min_size: Size) -> Self {
-//     let size = winit::dpi::LogicalSize::new(min_size.width, min_size.height);
-//     self.inner_builder = self.inner_builder.with_min_inner_size(size);
-//     self
-//   }
-
-//   /// Sets a maximum dimension size for the window.
-//   #[inline]
-//   pub fn with_max_inner_size(mut self, max_size: Size) -> Self {
-//     let size = winit::dpi::LogicalSize::new(max_size.width, max_size.height);
-//     self.inner_builder = self.inner_builder.with_max_inner_size(size);
-//     self
-//   }
-
-//   /// Sets a desired initial position for the window.
-//   #[inline]
-//   pub fn with_position(mut self, position: Point) -> Self {
-//     let position = winit::dpi::LogicalPosition::new(position.x, position.y);
-//     self.inner_builder = self.inner_builder.with_position(position);
-//     self
-//   }
-
-//   /// Sets whether the window is resizable or not.
-//   #[inline]
-//   pub fn with_resizable(mut self, resizable: bool) -> Self {
-//     self.inner_builder = self.inner_builder.with_resizable(resizable);
-//     self
-//   }
-
-//   /// Requests a specific title for the window.
-//   #[inline]
-//   pub fn with_title<T: Into<String>>(mut self, title: T) -> Self {
-//     self.inner_builder = self.inner_builder.with_title(title);
-//     self
-//   }
-
-//   /// Requests maximized mode.
-//   #[inline]
-//   pub fn with_maximized(mut self, maximized: bool) -> Self {
-//     self.inner_builder = self.inner_builder.with_maximized(maximized);
-//     self
-//   }
-
-//   /// Sets whether the window will be initially hidden or visible.
-//   #[inline]
-//   pub fn with_visible(mut self, visible: bool) -> Self {
-//     self.inner_builder = self.inner_builder.with_visible(visible);
-//     self
-//   }
-
-//   /// Sets whether the background of the window should be transparent.
-//   #[inline]
-//   pub fn with_transparent(mut self, transparent: bool) -> Self {
-//     self.inner_builder = self.inner_builder.with_transparent(transparent);
-//     self
-//   }
-
-//   /// Sets whether the window should have a border, a title bar, etc.
-//   #[inline]
-//   pub fn with_decorations(mut self, decorations: bool) -> Self {
-//     self.inner_builder = self.inner_builder.with_decorations(decorations);
-//     self
-//   }
-
-//   // /// Sets the window icon.
-//   // #[inline]
-//   // pub fn with_window_icon(mut self, window_icon:
-// Option<winit::window::Icon>)   // -> Self {   self.inner_builder =
-//   // self.inner_builder.with_window_icon(window_icon);   self
-//   // }
-// }
 
 /// A rx scheduler pool that block until all task finished before every frame
 /// end.
