+++ tests/animations_syntax_test.rs
@@ -1,4 +1,5 @@
 use ribir::prelude::*;
+use ribir_core::test::MockPointerId;
 use std::{cell::Cell, rc::Rc, time::Duration};
 // use winit::event::{DeviceId, MouseScrollDelta, TouchPhase, WindowEvent};
